+++ example/lib/src/album/model/photo.dart
@@ -1,0 +1,22 @@
+class Photo {
+  final int id;
+  final String title;
+  final String url;
+  final String thumbnailUrl;
+
+  Photo({
+    this.id,
+    this.title,
+    this.url,
+    this.thumbnailUrl,
+  });
+
+  factory Photo.fromJson(Map<String, dynamic> json) {
+    return Photo(
+      id: json['id'] as int,
+      title: json['title'] as String,
+      url: json['url'] as String,
+      thumbnailUrl: json['thumbnailUrl'] as String,
+    );
+  }
+}