+++ example/lib/src/album/ui/photos_list_empty.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class PhotosListEmpty extends StatelessWidget {
+  PhotosListEmpty({Key key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) => Center(child: Text('No photos found'));
+}