+++ example/lib/src/common/url.dart
@@ -1,0 +1,1 @@
+const baseUrl = 'https://jsonplaceholder.typicode.com';