+++ test/src/util/bdd.dart
@@ -1,0 +1,5 @@
+void given(void Function() given) => given();
+
+void when(void Function() when) => when();
+
+void then(void Function() then) => then();